refactor(app): add pages folders, update routes
