+++ src/components/main/Footer.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+export const Footer = () => {
+  // TODO
+  return <footer></footer>;
+};
