+++ src/components/main/Header.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+export const Header = () => {
+  // TODO
+  return <header></header>;
+};
