+++ src/components/pages/HomePage.js
@@ -1,19 +1,18 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import logo from '../logo.svg';
+import logo from '../../images/logo.svg';
 
-function HomePage() {
+export const HomePage = () => {
+  // TODO: rework this home page completely. Search field to lookup profile should go here
   return (
     <div>
       <img src={logo} className="App-logo" alt="logo" />
       <p>
         Edit <code>src/App.js</code> and save to reload.
       </p>
-      <Link className="App-link" to="/d3">
+      <Link className="App-link" to="/profile/:davidholyko">
         Learn React
       </Link>
     </div>
   );
-}
-
-export default HomePage;
+};
