+++ src/components/pages/index.js
@@ -1,0 +1,3 @@
+export * from './HomePage';
+export * from './RepoPage';
+export * from './UserPage';
