+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import App from './components/App';
-import './index.scss';
+import './css/index.scss';
 
 ReactDOM.render(
   <BrowserRouter>
